perf(sources filters): use 64bits numbers for points coordinates instead of 32bits

+++ Sources/Filters/Sources/CylinderSource/index.js
@@ -186,7 +186,7 @@ const DEFAULT_VALUES = {
   center: [0, 0, 0],
   direction: [0.0, 1.0, 0.0],
   capping: true,
-  pointType: 'Float32Array',
+  pointType: 'Float64Array',
 };
 
 // ----------------------------------------------------------------------------
